https://localorbit.atlassian.net/browse/LO-996
Lists of suborders for each seller that are grouped by market on the
'Record Payments to Suppliers' (vendor_payments) are now loaded in via
ajax. This leads to > 80% performance improvement when loading Ohio Valley Food Connection's
payments going back to Dec 2014.

financials.js.coffee is refactored to vanilla js, and updated to accomodate the ajax
loading and delegated event binding.